+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
-## [2022-05-23]
+## 0.2
+
+- Add command to add a Raycast Preference
+- Add command to add a Raycast command
+
+## 0.1
 
 - Initial release